Graficos se importan dinamicos

+++ api/utils/config.py
@@ -1,6 +1,6 @@
 import os
 import sys
-# from .. import parsers
+import json
 def singleton(cls):
     instance = None
     def getinstance(*args, **kwargs):
@@ -50,28 +50,46 @@ class Config:
         self.db_count_id=0 #para saber por que id voy generando los graficos
 
     def set_db_path(self,path):
-        file = open(path, "w")
-        file.write("[]")
-        file.close()
+        ''' Set 'path' as the data base path, a file where we puts the necesary info for our system
+        if the file exist then update the 'db_count_id'
+        if not then it will be create '''
+        if not os.path.isfile(path):
+            file = open(path, "w")
+            file.write("[]")
+            file.close()
+            self.db_count_id=0
+        else:
+            db_file = open(path, 'r')
+            info = db_file.read()
+            deserialization = json.loads(info)
+            self.db_count_id=0
+            for item in deserialization:
+                if item['id']>self.db_count_id:
+                    self.db_count_id=item['id']
+            self.db_count_id+=1
         self.db_path = path
-        self.db_count_id=0
 
-    def set_in_path(self, in_path='./data_generator/'):
+    def set_in_path(self, in_path):
+        ''' Set 'in_path' as the path where the program would be looking the input files '''
         self.input_path = os.path.abspath(in_path)
 
-    def set_out_path(self, out_path='./out'):
+    def set_out_path(self, out_path):
+        ''' Set 'out_path' as the path where the program would put the output files '''
         self.output_path = os.path.abspath(out_path)
 
     def set_output_name(self, output_name):
-        """ update self.output_name to the params from the input  """
+        """ set output_name as the default name for the outputs files
+        and output count as the number of the file that we have for not repetitions"""
         self.output_name = output_name if output_name != None else "out_file"
         self.output_count=len(os.listdir(Config().output_path))
 
     def set_parsers_path(self,p_path):
+        ''' Set 'p_path' as the path where the program would be looking the parsers '''
         self.parsers_path = os.path.abspath(p_path)
         self.__set_parsers()
 
     def set_graphs_path(self,g_path):
+        ''' Set 'g_path' as the path where the program would be looking the graphs '''
         self.graphs_path=os.path.abspath(g_path)
         self.__set_graphs()
 
@@ -89,32 +107,48 @@ class Config:
                 import_module = compile("import "+item[:-3], 'e', mode='exec')
                 exec(import_module)
                 # temp = eval(str(class_name+"()"))
-                temp = eval(str(item[:-3]+"."+class_name+"()"))
-                if dir(temp).__contains__("parse") and dir(temp).__contains__("help"):
-                    self.parsers.append(item[:-3])
-                else:
+                try:
+                    temp = eval(str(item[:-3]+"."+class_name+"()"))
+                    if dir(temp).__contains__("parse") and dir(temp).__contains__("help"):
+                        self.parsers.append(item[:-3])
+                    else:
+                        print("WARNING-- THE FILE "+item +
+                          " DOES NOT IMPLEMENT THE INTERFACE")
+                except:
                     print("WARNING-- THE FILE "+item +
                           " DOES NOT IMPLEMENT THE INTERFACE")
 
     def __set_graphs(self):
         ''' Update self.graphs to a list with all graphs who are implemented '''
+        sys.path.append(self.graphs_path)
         self.graphs = []
         for item in os.listdir(self.graphs_path):
             p = self.graphs_path+"/"+item
             if (not os.path.isdir(p)):
-                self.graphs.append(item[:-3])
+                class_name = self.get_parser_class_name(item[:-3])
+                import_module = compile("import "+item[:-3], 'e', mode='exec')
+                exec(import_module)
+                try:
+                    temp = eval(str(item[:-3]+"."+class_name+"()"))
+                    if dir(temp).__contains__("graphic") and dir(temp).__contains__("type"):
+                        self.graphs.append(item[:-3])
+                    else:
+                        print("WARNING-- THE FILE "+item +
+                          " DOES NOT IMPLEMENT THE INTERFACE")
+                except:
+                    print("WARNING-- THE FILE "+item +
+                          " DOES NOT IMPLEMENT THE INTERFACE")
+
 
     def set_available_parsers(self,in_parsers='all'):
         """ Update self.available_parsers to a list of parsers
             who are going to be plot in depends of the arguments of the input"""
         available_parsers = in_parsers.split(',')
         self.available_parsers = []
-
         if in_parsers != "all":
             for parser in available_parsers:
                 if self.parsers.__contains__(parser):
-                    self.available_parsers.append(
-                        self.get_parser_class_name(parser))
+                    self.available_parsers.append(self.get_parser_class_name(parser))
                 else:
                     print("WARNING-- THE PARSER "+parser + " IS NOT DEFIND")
         else:
@@ -126,15 +160,15 @@ class Config:
         graphs  who are going to be plot in depends of the arguments of the input"""
         available_graphs = in_graphs.split(',')
         self.available_graphs = []
-
         if in_graphs != "all":
             for graph in available_graphs:
                 if self.graphs.__contains__(graph):
-                    self.available_graphs.append(graph)
+                    self.available_graphs.append(self.get_parser_class_name(graph))
                 else:
                     print("WARNING-- THE GRAPH "+graph + " IS NOT DEFIND")
         else:
-            self.available_graphs = self.graphs
+            self.available_graphs = [self.get_parser_class_name(
+                graph) for graph in self.graphs]
 
     def set_parsers_help(self,in_parsers_help):
         """ Update self.prarsers_help to a list of parsers who's help are going to be show"""
